Refactor project structure to accommodate persistence layer (contains: model, dto, converter, repository)
