+++ src/main/java/com/example/echo_api/service/auth/AuthService.java
@@ -1,9 +1,9 @@
 package com.example.echo_api.service.auth;
 
-import com.example.echo_api.dto.request.SignInRequest;
-import com.example.echo_api.dto.request.SignUpReqest;
 import com.example.echo_api.exception.user.UserAlreadyExistsException;
 import com.example.echo_api.exception.user.UserNotFoundException;
+import com.example.echo_api.persistence.dto.request.SignInRequest;
+import com.example.echo_api.persistence.dto.request.SignUpReqest;
 
 public interface AuthService {
